requirements.txt
pylint
twitter-to-sqlite
tqdm
sqlitebiter
datasette
sqlite-utils
sqlalchemy
urlextract
idna
uritools
appdirs
requests
furl
python-slugify
tldextract
pyyaml
bonobo
ucache
peewee
black